+++ lib/screens/homepage/app_group.dart
@@ -4,11 +4,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_quiver/screens/connect4/connect4.dart';
 import 'package:flutter_quiver/screens/ct_suppliers_backlog/ct_suppliers_backlog.dart';
 import 'package:flutter_quiver/screens/demand_bids/demand_bids.dart';
+import 'package:flutter_quiver/screens/examples/daterange_example.dart';
 import 'package:flutter_quiver/screens/examples/dropdown_example.dart';
 import 'package:flutter_quiver/screens/examples/expansion_panel.dart';
 import 'package:flutter_quiver/screens/examples/inherited_widget_example.dart';
 import 'package:flutter_quiver/screens/examples/multiselect_menu_button_example.dart';
-import 'package:flutter_quiver/screens/examples/two_linked_multiselects_example.dart';
+import 'package:flutter_quiver/screens/examples/linked_multiselects_example.dart';
 import 'package:flutter_quiver/screens/historical_gas/historical_gas_ui.dart';
 import 'package:flutter_quiver/screens/historical_option_pricing/historical_option_pricing_ui.dart';
 import 'package:flutter_quiver/screens/hourly_shape/hourly_shape.dart';
@@ -74,6 +75,7 @@ class AppGroup extends StatefulWidget {
       MenuItem(url: Weather.route, title: 'Weather'),
     ],
     'Examples': [
+      MenuItem(title: 'Date range picker', url: DateRangePickerExample.route),
       MenuItem(title: 'Dropdown without lag', url: DropdownExample.route),
       MenuItem(title: 'Expansion panel', url: ExpansionPanelExample.route),
       MenuItem(
@@ -82,13 +84,13 @@ class AppGroup extends StatefulWidget {
           title: 'Multiselect dropdown',
           url: MultiSelectMenuButtonExample.route),
       MenuItem(
-          title: 'Two linked multiselect dropdowns',
-          url: TwoLinkedMultiSelectsExample.route),
+          title: 'Linked multiselect dropdowns',
+          url: LinkedMultiSelectsExample.route),
     ],
   };
 
   @override
-  _AppGroupState createState() => _AppGroupState();
+  State<AppGroup> createState() => _AppGroupState();
 }
 
 class _AppGroupState extends State<AppGroup> {